requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
python-multipart==0.0.6

# Database (uncomment when needed)
# supabase==2.0.3
# psycopg2-binary==2.9.9
# sqlalchemy==2.0.23

# AI/ML (add when implementing AI features)
# openai==1.3.0
# langchain==0.0.340
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
